Image uploads complete
Images will now be deleted from Supabase Storage when a project creation is cancelled, when a new image is uploaded or when the project is deleted.

Project ID now uses uuid to enable coupling of projects and thumbnail images. Had to delete all existing projects to do this but should not need to in future.

Limited the initial intake of project data to what is displayed on the project card.

Project card now displays completion status and time since project was created.